fluid images

+++ flask_app/controllers/instruments.py
@@ -13,7 +13,7 @@ def allowed_file(filename):
 # Create Instruments Controller
 @app.route('/instruments/new', methods=['POST', 'GET'])
 def create_instrument():
-    # if 'user_id' not in session: return redirect('/')
+    if 'user_id' not in session: return redirect('/') # *** route protection, must be logged in
     if request.method == 'POST':
         # create a new instrument and retrieve the id
         instrument_id = instrument.Instrument.create_new_instrument(request.form)
@@ -32,7 +32,7 @@ def create_instrument():
 # Read Instruments Controller
 @app.route('/home')
 def show_user_instruments():
-    # if 'user_id' not in session: return redirect('/')
+    if 'user_id' not in session: return redirect('/') # ***
     all_instruments = instrument.Instrument.get_all_instruments_with_users()
     # REPLACE THIS WITH A QUERY!!!!
     owned_instruments = [inst for inst in all_instruments if inst.seller_id == session['user_id']]
@@ -43,7 +43,7 @@ def show_user_instruments():
 
 @app.route('/instruments/all')
 def show_all_instruments():
-    # if 'user_id' not in session: return redirect('/')
+    if 'user_id' not in session: return redirect('/') # ***
     all_instruments = instrument.Instrument.get_all_instruments_with_users()
     # REPLACE THIS WITH A QUERY!!!
     instruments_for_sale = [inst for inst in all_instruments if inst.sold == False]
@@ -51,7 +51,7 @@ def show_all_instruments():
 
 @app.route('/instruments/<int:id>')
 def instrument_card(id):
-    # if 'user_id' not in session: return redirect('/')
+    if 'user_id' not in session: return redirect('/') # ***
     this_instrument = instrument.Instrument.get_instrument_by_id(id)
     return render_template('one_instrument.html', instrument = this_instrument)
 
@@ -64,11 +64,12 @@ def get_instrument_image(filename):
 # Update Instruments Controller
 @app.route('/instruments/edit/<int:id>', methods=['POST', 'GET'])
 def edit_instrument(id):
-    # if 'user_id' not in session: return redirect('/')
+    if 'user_id' not in session: return redirect('/') # ***
     if request.method == 'POST':
         updated = instrument.Instrument.update_instrument(request.form)
         if updated:
             return redirect('/home')
+    # if 'GET' request:
     instrument_to_update = instrument.Instrument.get_instrument_by_id(id)
     if instrument_to_update.user_id == session['user_id']:
         inst_select = instrument.Instrument.get_instrument_select()
@@ -76,11 +77,11 @@ def edit_instrument(id):
         return render_template('edit_instrument.html', instrument = instrument_to_update, 
                             inst_select = inst_select, quality_select = quality_select)
     else:
-        return redirect('/users/logout')
+        return redirect('/users/logout') # hacking protection
     
 @app.route('/instruments/purchase/<int:id>', methods=['POST'])
 def purchase_instrument(id):
-    # if 'user_id' not in session: return redirect('/')
+    if 'user_id' not in session: return redirect('/') # ***
     instrument.Instrument.change_owner(id)
     return redirect('/home')
 
@@ -88,8 +89,8 @@ def purchase_instrument(id):
 # Delete Instruments Controller
 @app.route('/instruments/delete/<int:id>')
 def delete_instrument(id):
-    # if 'user_id' not in session: return redirect('/')
+    if 'user_id' not in session: return redirect('/') # ***
     if instrument.Instrument.delete_instrument_by_id(id):
         return redirect('/home')
     else:
-        return redirect('/users/logout')
+        return redirect('/users/logout') # hacking protection
